Add --device support for Windows
Adds support for --device in Windows. This must take the form of:
--device='class/{clsid}'
